Save player rank, not rank points.

+++ WowPacketParser/Store/Objects/UpdateFields/IPlayerData.cs
@@ -10,7 +10,7 @@ namespace WowPacketParser.Store.Objects.UpdateFields
         uint PlayerBytes1 { get; }
         uint PlayerBytes2 { get; }
         uint PlayerFlags { get; }
-        uint PvPRank { get; }
+        byte PvPRank { get; }
         IVisibleItem[] VisibleItems { get; }
         uint GuildRankID { get; }
         ChrCustomizationChoice[] GetCustomizations();
